+++ lib/abstract_session.dart
@@ -17,7 +17,6 @@
  * along with FFmpegKit.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'ffmpeg_kit_config.dart';
@@ -28,6 +27,7 @@ import 'log_callback.dart';
 import 'log_redirection_strategy.dart';
 import 'media_information.dart';
 import 'media_information_session.dart';
+import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'return_code.dart';
 import 'session.dart';
 import 'session_state.dart';