+++ lib/ffmpeg_kit_config.dart
@@ -17,7 +17,6 @@
  * along with FFmpegKit.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'ffmpeg_session.dart';
@@ -29,6 +28,7 @@ import 'log_callback.dart';
 import 'log_redirection_strategy.dart';
 import 'media_information_session.dart';
 import 'media_information_session_complete_callback.dart';
+import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'session.dart';
 import 'session_state.dart';
 import 'signal.dart';