+++ lib/ffmpeg_session.dart
@@ -17,7 +17,6 @@
  * along with FFmpegKit.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'abstract_session.dart';
@@ -25,6 +24,7 @@ import 'ffmpeg_kit_config.dart';
 import 'ffmpeg_session_complete_callback.dart';
 import 'log_callback.dart';
 import 'log_redirection_strategy.dart';
+import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'src/ffmpeg_kit_factory.dart';
 import 'statistics.dart';
 import 'statistics_callback.dart';