+++ lib/ffprobe_kit.dart
@@ -17,7 +17,6 @@
  * along with FFmpegKit.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'ffmpeg_kit_config.dart';
@@ -26,6 +25,7 @@ import 'ffprobe_session_complete_callback.dart';
 import 'log_callback.dart';
 import 'media_information_session.dart';
 import 'media_information_session_complete_callback.dart';
+import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'src/ffmpeg_kit_factory.dart';
 
 /// Main class to run "FFprobe" commands.