+++ lib/packages.dart
@@ -17,10 +17,10 @@
  * along with FFmpegKit.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'ffmpeg_kit_config.dart';
+import 'platform_interface/ffmpeg_kit_flutter_platform_interface.dart';
 
 /// Helper class to extract binary package information.
 class Packages {