+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3
+
+* 
+* Fixed Android and MacOS bindings, Support 16kb page sizes
+* Upgraded FFmpegKitFlutterPlugin.java to work with Flutter 3.32.4
+
+
 ## 1.0.2
 
 * Initial release